Updated README.md
Added what still needs to be implemented in the most current update and will publish it under Update 2.0 once implemented. Need to implement gh-pages branch to this site as well as each of the other sites. Test to see if there is a performance difference between Blue Host and Digital Ocean servers.